Require user to specify User entity class
Following FOSUserBundle, the User entity in this bundle will serve as
the mapped super class. This means users will need to define their own
User entities and specify the FQCN in this bundle's configuration; same
as FOSUserBundle.

This allows us to get rid of the nasty hack to guess the class name of
the entity being used in the given Symfony app.

Fixes #1